Fix empty routing key in the headers (#27)
# Overview

Earlier, we were using local variables to store the state in addition to
Chrome's Local Storage. This was problematic wherein the service worker
(background processes) idles out and loses state.

Under that situation, when the user accesses the extension, two things
happen:
- The background process sets the local state from local storage
- Update dynamic roles to set the header

There's a possibility of a race here wherein it sets the dynamic rules
to set headers from memory which hasn't yet been reinstated from LS.

As a solution, this PR lets go of the memory vars altogether, and sets
the dynamic rules with the values from LocalStorage.